Exclusively use int64 for internal arithmetic.
Some of the constants in use (and some supported int values) overflow
int32, which cauees the build to fail on 32-bit systems.

To avoid changing any external-facing APIs, this commit changes only
internal, transient variables used in datetime arithmetic. That leaves a
few places (such as the "microseconds" argument to New) where integer
values outside the 32 bit range would work, but just won't be available
on 32-bit machines.

To test this on a 64 bit machine, run:

    GOARM=6 GOARCH=arm GOOS=linux go test

(this should fail on the exec, not on compilation)

Fixes #6.